vulnops-intelligence-platform/backend/requirements/base.txt
alembic==1.18.4
    # via -r requirements/base.in
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   starlette
    #   watchfiles
bcrypt==4.3.0
    # via -r requirements/base.in
click==8.3.2
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via -r requirements/base.in
fastapi==0.115.14
    # via -r requirements/base.in
greenlet==3.4.0
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   email-validator
joblib==1.5.3
    # via
    #   -r requirements/base.in
    #   scikit-learn
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
numpy==2.4.4
    # via
    #   -r requirements/base.in
    #   pandas
    #   scikit-learn
    #   scipy
pandas==2.3.3
    # via -r requirements/base.in
prometheus-client==0.25.0
    # via prometheus-fastapi-instrumentator
prometheus-fastapi-instrumentator==7.1.0
    # via -r requirements/base.in
psycopg==3.3.3
    # via -r requirements/base.in
psycopg-binary==3.3.3
    # via psycopg
pydantic==2.12.5
    # via
    #   -r requirements/base.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via -r requirements/base.in
pyjwt==2.12.1
    # via -r requirements/base.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.26
    # via -r requirements/base.in
pytz==2026.1.post1
    # via pandas
pyyaml==6.0.3
    # via uvicorn
scikit-learn==1.8.0
    # via -r requirements/base.in
scipy==1.17.1
    # via scikit-learn
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.49
    # via
    #   -r requirements/base.in
    #   alembic
starlette==0.46.2
    # via
    #   fastapi
    #   prometheus-fastapi-instrumentator
threadpoolctl==3.6.0
    # via scikit-learn
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2026.1
    # via
    #   pandas
    #   psycopg
uvicorn==0.30.6
    # via -r requirements/base.in
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
